+++ src/query/sql/src/planner/optimizer/rule/rewrite/filter_join/outer_join_to_inner_join.rs
